% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilesize.R
\name{geom_tile_size}
\alias{geom_tile_size}
\alias{stat_tile_size}
\title{A version of geom_tile with variable cell size}
\usage{
geom_tile_size(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  sizedims = "xy",
  sizelims = c(0, 1),
  ...
)

stat_tile_size(
  mapping = NULL,
  data = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  sizedims = "xy",
  sizelims = c(0, 1),
  ...
)
}
\arguments{
\item{data, mapping, na.rm, show.legend, inherit_aes}{Standard ggplot2
parameters common across geoms and stats.}

\item{sizedims}{String specifying which dimensions to scale: either "xy"
(default), "x", or "y".}

\item{sizelims}{Length-2 numeric vector specifying the range of relative
cell widths used to represent the `size` variable. Under the default,
c(0, 1), the smallest value will have an area of zero and the largest
will occupy a full cell space. Values greater than 1 are allowed.}

\item{...}{Other arguments passed on to `layer()`.}
}
\description{
This stat is analogous to geom_tile or geom_raster, but with an additional
aesthetic parameter that controls the size of the tile.
}
\section{Functions}{
\itemize{
\item \code{stat_tile_size()}: Stat corresponding to geom_tile_size

}}
